requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
google-generativeai
thefuzz
python-Levenshtein
httpx
pytest
requests
